StockLens - W/server/requirements.txt
flask
flask-cors
pymongo
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt
requests
tenacity
